Hook up to xstate store and set initial state
