+++ workspaces/ui-v2/src/optic-components/pages/diffs/ReviewEndpointDiffPage/index.ts
@@ -1,0 +1,1 @@
+export * from './ReviewEndpointDiffPage';
